backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
matplotlib==3.4.3
seaborn==0.11.2
python-pptx==0.6.21
openpyxl==3.0.9
torch==2.1.0
ultralytics==8.0.196
pyyaml==6.0.1